Added basic calendar function.
The calendar allows you to view which threads are started
or ended on a given day. This is all lead-up work really,
the next task is to allow any thread to be associated with
an 'event-date', which we're already defaulting the
calendar view to.

The default calendar view at the moment thus does
nothing but show an empty calendar, it's all poised
and ready to show topics with associated events though.

The calendar itself is a bunch of CSS boxes in <li> tags
with appropriate overflow and padding boxes. Thus, no
tables and we can put each day's stuff into a single
phpbb variable array without having to worry about
which row and column things should be in. It's
pretty neat.